+++ apdt/ml/__init__.py
@@ -10,3 +10,4 @@ from apdt.ml.general import DataSet, TFModel
 from apdt.ml.tf_model import wavenet_weight, WaveNet, mlp_weight, MLP, lstm_weight, LSTM
 from apdt.ml.tf_model import Transformer_ar_decoder_weight, Transformer_ar_decoder
 from apdt.ml.tf_model import _WaveNet
+from apdt.ml.utils import stacked_window